+++ podchat/src/main/java/com/fanap/podchat/persistance/AppDatabase.java
@@ -12,7 +12,7 @@ import com.fanap.podchat.persistance.dao.MessageDao;
 @Database(entities = {Contact.class}, version = 1, exportSchema = false)
 public abstract class AppDatabase extends RoomDatabase {
 
-    public static final String DATABASE_DB = "cache.db";
+    public static final String DATABASE_DB = "isCacheable.db";
     private static AppDatabase appDatabase;
 
     public static AppDatabase getInstance(Context context) {